added manage page and its functions for users listings

+++ app/Http/Controllers/ListingController.php
@@ -61,6 +61,12 @@ class ListingController extends Controller
     }
 
     public function update(Request $request,Listing $listing){
+         
+        // Make sure logged in user is owner
+        if($listing->user_id != auth()->id){
+            abort(403,'Unauthorized Action');
+        }
+            
         $formFields = $request->validate([
              'title' => 'required',
              'company' => ['required'],
@@ -83,8 +89,17 @@ class ListingController extends Controller
 
     // Delete  Listing
     public function destroy(Listing $listing){
+        // Make sure logged in user is owner
+        if($listing->user_id != auth()->id){
+            abort(403,'Unauthorized Action');
+        }
             $listing->delete();
             return redirect('/')->with('message','listing deleted successfully');
+    }
+
+    // Manage Listings
+    public function manage(){
+        return view('listings.manage',['listings' => auth()->user()->listings()->get()]);
     }
 
 }
